requirements.txt
# Serial ↔ MQTT bridge (serial_to_mqtt.py)
pyserial>=3.5
paho-mqtt>=1.6.1
